from_mqtt/mqtt_sql/XMLtoSQL/requirements.txt
logging==0.4.9.6
pymssql==2.3.0
PyYAML==6.0.1